# backend/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: klaxoon-bookmarks
  description: A REST API to manage bookmarks

consumes:
- application/json
produces:
- application/json

definitions:
  Bookmark:
    type: object

    required:
      - url
      - title
      - author
      - createDate
    properties:
      url:
        type: string
      title:
        type: string
      author:
        type: string
      createDate:
        type: string
      width:
        type: string
      height:
        type: string
      duration:
        type: integer

  Bookmarks:
    type: array
    items:
      $ref: "#/definitions/Bookmark"

  keyWords:
    type: array
    items:
      type: string

  bookmarkUrl:
    type: object
    required:
      - bookmarkUrl
    properties:
      bookmarkUrl:
        type: string

  Error:
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string

paths:
  /bookmarks:
    x-swagger-router-controller: controller
    get:
      description: get all bookmarks
      operationId: getBookmarks
      responses:
        200:
          description: An array of bookmarks
          schema:
            $ref: "#/definitions/Bookmarks"
        default:
          description: unexpected error
          schema:
            $ref: "#/definitions/Error"

    post:
      description: Create a bookmark
      operationId: createBookmark
      parameters:
      - name: body
        in: body
        schema:
          $ref: "#/definitions/bookmarkUrl"
      responses:
        201:
          description: Null response
        default:
          description: unexpected error
          schema:
            $ref: "#/definitions/Error"

  /bookmarks/{type}/{id}:
    x-swagger-router-controller: controller
    get:
      description: get given bookmark
      operationId: getBookmark
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the bookmark
          type: string
        - name: type
          in: path
          required: true
          description: The type of the bookmark (video or photo)
          type: string
          enum: [video, photo]

      responses:
        200:
          description: Expected response to a valid request
          schema:
            $ref: "#/definitions/Bookmark"
        default:
          description: unexpected error
          schema:
            $ref: "#/definitions/Error"

    put:
      description: update given bookmark
      operationId: updateBookmark
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the bookmark
          type: string

        - name: type
          in: path
          required: true
          description: The type of the bookmark (video or photo)
          type: string
          enum: [video, photo]

        - name: body
          in: body
          schema:
            $ref: "#/definitions/keyWords"

      responses:
        204:
          description: No content response
        default:
          description: unexpected error
          schema:
            $ref: "#/definitions/Error"
  
    delete:
      description: delete given bookmark
      operationId: deleteBookmark
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the bookmark
          type: string
        - name: type
          in: path
          required: true
          description: The type of the bookmark (video or photo)
          type: string
          enum: [video, photo]
      responses:
        204:
          description: No content response
        default:
          description: unexpected error
          schema:
            $ref: "#/definitions/Error"
  /:
    x-swagger-router-controller: controller
    get:
      operationId: home
      responses:
        default:
          description: Returns hello
          schema:
            type: string